% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkaryo.R
\name{bin_track}
\alias{bin_track}
\title{Bins a track based on provided bin size and step.
Regions from the track are assigned to the bins when they are completely
include ('within' method) or overlap even partially ('overlap' method).}
\usage{
bin_track(track, size, step, method = "within", fun.aggreg = mean, ...)
}
\arguments{
\item{track}{(data.table) BED5+ track data table}

\item{size}{(numeric) bin size in nt}

\item{step}{(numeric) bin step in nt}

\item{method}{(string) either 'within' or 'overlap'}

\item{fun.aggreg}{(function) how to aggregate values in bins}

\item{...}{(mixed) additional parameters to pass to fun.aggreg}
}
\description{
Bins a track based on provided bin size and step.
Regions from the track are assigned to the bins when they are completely
include ('within' method) or overlap even partially ('overlap' method).
}
