% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkaryo.R
\docType{class}
\name{ggkaryo}
\alias{ggkaryo}
\title{ggkaryo: a class for karyotype plotting and overlaying.}
\description{
Lorem ipsum dolor sit amet, consectetur adipisicing elit, sed do eiusmod
tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,
quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo
consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse
cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non
proident, sunt in culpa qui officia deserunt mollit anim id est laborum.
}
\section{Fields}{

\describe{
\item{\code{n_chrom}}{(numerical) number of chromosomes, default: 24}

\item{\code{hetero}}{(character) heterosome labels (without "chr"),
default: c("X", "Y")}

\item{\code{chrom_width}}{(numerical) width of the ideograms, default: 1}

\item{\code{chrom_padding}}{(numerical) space between ideograms, default: 5}

\item{\code{track_palette_name}}{(character) name of RColorBrewer palette for track}

\item{\code{giemsa_palette}}{(character) vector of colors for the giemsa bands}

\item{\code{giemsa_levels}}{(character) vector of giemsa band levels}

\item{\code{opposite}}{(logical) to plot profiles on both sides of the ideogram}

\item{\code{data}}{(list) contains ggkaryo data for plotting}
}}

\section{Methods}{

\describe{
\item{\code{add_arm_boxes()}}{Adds boxes around chromosome arms.}

\item{\code{add_chrom_labels()}}{Adds chromosome labels.}

\item{\code{add_lois(data, position, colorName, alpha = 1)}}{docline}

\item{\code{add_track(track, step, position = "auto", color = "auto", alpha = 0.5)}}{Adds a profile to the current ggkaryo plot. The input track must have
already been binned with a consistent step. A consistent step is needed
to automatically set any gap to 0 in the profile.
See \code{prepare_track} for more details.}

\item{\code{chrom2id(chrom)}}{Converts a chromosome signature (seqname) to a numerical id.
\describe{
  \item{\code{chrom}}{
    (string) chromosome signature (e.g., 'chr1' or '1')}
  \item{returns}{numeric: chromosome numerical ID}
}}

\item{\code{chromID2x(chromID)}}{Retrieve the position of a chromosome on the X axis.
\describe{
  \item{\code{chromID}}{(numeric)}
  \item{returns}{numeric: chromosome position on the X axis}
}}

\item{\code{get_color(color, trackID)}}{Extracts, in order, track colors from the ggk$track_palette_name.
See RColorBrewer for more details.
\describe{
  \item{\code{color}}{(character) a color or 'auto'}
  \item{\code{trackID}}{(numeric) track number}
}}

\item{\code{get_next_position(position)}}{Selects position for the next track in such a fashion to balance out
left/right sides of the ideograms.
\describe{
  \item{\code{position}}{(character) 'right', 'left', or 'auto'}
}}

\item{\code{initialize(giemsa, ..., n_chrom = 24, hetero = c("X", "Y"),
  chrom_width = 1, chrom_padding = 5, track_palette_name = "Paired",
  giemsa_palette = c("#DDDDDD", "#9A9A9A", "#787878", "#555555", "#333333",
  "#FF0000", "#C4FFFC", "#AFE6FF"), giemsa_levels = c("gneg", "gpos25",
  "gpos50", "gpos75", "gpos100", "acen", "gvar", "stalk"), opposite = F)}}{Initializer method. See \code{ggkaryo} description for more details}

\item{\code{norm2x(chromID, norm, position)}}{Converts normalized score to X coordinate in the ggkaryo plot.
\describe{
  \item{\code{chromID}}{(numeric)}
  \item{\code{norm}}{(numeric) normalized score}
  \item{\code{position}}{(character) 'left' or 'right'}
  \item{returns}{numeric: normalized score X coordinate}
}}

\item{\code{plot_base()}}{Plots the current ggkaryo object (only basic layers).}

\item{\code{plot_full()}}{Plots the current ggkaryo object with tracks and lois.}

\item{\code{prep4bands()}}{Prepares data for plotting chromosome bands.
Builds .self$data[['bands']] object.}

\item{\code{prep4boxes()}}{Prepares data for plotting chromosome arm boxes. Chromosome arms are
identified based on the 'acen' bands that are used to divide each
chromosomes in two arms. Builds .self$data[['boxes']] object.}

\item{\code{prep4karyo(giemsa)}}{Builds a data.table to plot the ideograms and (optionally) boxes around
each chromosome arm.
\describe{
  \item{\code{giemsa}}{(character) path to giemsa BED5+ file}
  \item{\code{giemsa}}{(data.table) data table with giemsa BED5+ data}
  \item{returns}{data.table: adjusted giemsa data.table}
}}

\item{\code{prep4labels()}}{Prepares data for plotting chromosome labels.
Builds .self$data[['chrom_labels']] object.}

\item{\code{prep4plot()}}{Prepares for plotting. Builds .self$data[['plot']] object.}

\item{\code{read_giemsa(giemsa)}}{Reads a Giemsa bed file. Adds chromID, bandID, and X columns.
\describe{
  \item{\code{giemsa}}{(character) path to giemsa BED5+ file}
  \item{\code{giemsa}}{(data.table) data table with giemsa BED5+ data}
  \item{returns}{data.table: adjusted giemsa data.table}
}}
}}

