% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giemsa-data.R
\docType{data}
\name{giemsa}
\alias{giemsa}
\title{Human genome (hg19) Giemsa bands obtained from UCSC.}
\format{A data table with 862 rows and 5 columns:
\describe{
  \item{V1}{chromosome}
  \item{V2}{start position}
  \item{V3}{end position}
  \item{V4}{chromosome band name}
  \item{V5}{giemsa band value}
}}
\source{
\url{https://genome.ucsc.edu/}
}
\usage{
giemsa
}
\description{
Human genome (hg19) Giemsa bands obtained from UCSC.
}
\keyword{datasets}
