% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lois-data.R
\docType{data}
\name{lois}
\alias{lois}
\title{Example dataset with loci of interest.}
\format{A data table with 3262 rows and 8 columns:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{center}{midpoint position}
  \item{gene_id}{ensemble gene ID}
  \item{sample}{NES/Neuron}
  \item{gene_name}{HUGO gene name}
  \item{annotation}{gene type}
}}
\usage{
lois
}
\description{
Example dataset with loci of interest.
}
\keyword{datasets}
