% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkaryo.R
\name{prepare_track}
\alias{prepare_track}
\title{Prepares a profile to be added to a ggkaryo plot. The input track must have
already been binned with a consistent step. A consistent step is needed
to automatically set any gap to 0 in the profile.}
\usage{
prepare_track(track, step, position = "auto", color = "auto",
  alpha = 0.5)
}
\arguments{
\item{track}{(string) path to BED5+ file.}

\item{step}{(numerical) bin step in nt.}

\item{position}{(string) one of auto|left|right. 'left' can be used only
if opposite=T was used when initializing the ggkaryo object.}

\item{color}{(string) either 'auto' or a color string.}

\item{alpha}{(numerical) opacity level.}

\item{track}{(data.table) BED5+ data table.}
}
\description{
Prepares a profile to be added to a ggkaryo plot. The input track must have
already been binned with a consistent step. A consistent step is needed
to automatically set any gap to 0 in the profile.
}
