% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-data.R
\docType{data}
\name{track}
\alias{track}
\title{Example track binned with size of 1 Mb and step of 100 kb.}
\format{A data table with 30318 rows and 5 columns:
\describe{
  \item{chrom}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{V4}{bin name}
  \item{value}{track bin value}
}}
\usage{
track
}
\description{
Example track binned with size of 1 Mb and step of 100 kb.
}
\keyword{datasets}
